Don't buildrequire texinfo-tex and save lots of build deps
texinfo-tex pulls many packages from the texlive distribution. If we can
avoid importing texlive (500 RPMs!) into our repositories just because a
package (devtoolset-11-gcc) which is itself a build dependency of other
packages needs it only to build texinfo pages, let's avoid it.

Fortunately, a local build succeeds without this heavy dependency, so
either it's a stale dependency, or the build scripts are smart enough to
do without it.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>